# groSCery-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "201 Project API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Login, Logout, and Registration"
                ],
                "summary": "Login a user",
                "operationId": "App\\Http\\Controllers\\UserController::login",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": {
                                        "token": "1KJGK234..."
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Login, Logout, and Registration"
                ],
                "summary": "Logout a user",
                "operationId": "App\\Http\\Controllers\\UserController::logout",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "Token has been revoked"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Login, Logout, and Registration"
                ],
                "summary": "Register a new user",
                "operationId": "App\\Http\\Controllers\\UserController::register",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "c_password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": {
                                        "token": "1KJGK234..."
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "409": {
                        "description": "User could not be created due to a conflict. User already exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "error": "User already exists"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/details": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a users basic information",
                "operationId": "App\\Http\\Controllers\\UserController::details",
                "responses": {
                    "200": {
                        "description": "User details have been retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "properties": {
                                                "id": {
                                                    "type": "int"
                                                },
                                                "group_id": {
                                                    "type": "int"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "pic_url": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": {
                                        "id": 2,
                                        "group_id": null,
                                        "name": "Kristof",
                                        "pic_url": null,
                                        "email": "kristof@usc.edu",
                                        "created_at": "2018-10-30 01:13:02"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/users/items": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a user's subscibed items",
                "operationId": "App\\Http\\Controllers\\UserController::items",
                "responses": {
                    "200": {
                        "description": "User items have been retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "properties": {
                                                "id": {
                                                    "type": "int"
                                                },
                                                "group_id": {
                                                    "type": "int"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "pic_url": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": [
                                        {
                                            "id": 1,
                                            "group_id": 1,
                                            "name": "Banana",
                                            "in_stock": 1,
                                            "pic_url": null
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        }
    },
    "components": {
        "responses": {
            "401": {
                "description": "Request is not authorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "error": "Unauthorized"
                        }
                    }
                }
            },
            "400": {
                "description": "Bad request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "error": "Bad request"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT"
            }
        }
    }
}